requirements.txt
fastapi==0.104.1
PyMuPDF==1.23.7
python-dotenv==1.0.0
Requests==2.31.0
httpx==0.25.2
python-multipart==0.0.6